requirements.txt
beautifulsoup4
requests
kafka-python
schedule
hdfs
pyspark
pyyaml
pandas
streamlit